// src/partner/external-operation.enums.ts
export enum ExternalOperationResourceType {
  WITHDRAWAL = 'WITHDRAWAL',
}

export enum ExternalOperationReferenceType {
  REQUEST = 'REQUEST',
  OPERATION = 'OPERATION',
  TRANSACTION = 'TRANSACTION',
  SETTLEMENT = 'SETTLEMENT',
  CALLBACK = 'CALLBACK',
  STATEMENT_ROW = 'STATEMENT_ROW',
  PROVIDER_IDEMPOTENCY = 'PROVIDER_IDEMPOTENCY',
}

export enum ExternalOperationReferenceSource {
  REQUEST = 'REQUEST',
  ACKNOWLEDGEMENT = 'ACKNOWLEDGEMENT',
  STATUS_QUERY = 'STATUS_QUERY',
  CALLBACK = 'CALLBACK',
  STATEMENT = 'STATEMENT',
  REPORT = 'REPORT',
}